requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
django==5.2.6
djangorestframework==3.16.1
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
sqlparse==0.5.3
djangorestframework==3.16.1
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.5
paho-mqtt==2.1.0